requirements.txt
fastapi
uvicorn
vnstock
xnoapi
apscheduler
python-dotenv
cachetools>=5.0.0
pandas-ta